+++ CHANGELOG.md
@@ -9,7 +9,7 @@ deploy webhooks to print changelog successfully
 
 That's all, matcher will stop when detects next line started with `###` match
 
-### 4.3.2 Release (23.01.2020 - 09.02.2020)
+### 4.3.2 Release (23.01.2020 - 02.03.2020)
 * Added new Zombie VillagerBuster (wave 15 and up)
 * Fixed zombie PlayerBuster & GolemBuster
 * Fixed spectator can hit zombie_buster zombies
@@ -26,6 +26,8 @@ That's all, matcher will stop when detects next line started with `###` match
       * mapname - name of arena map
 * Fixed mysql database - Create a new one and your stats should be saved correctly
 * Fixed inventory bug
+* Fixed incompatibility with UltimateStacker and similar 
+* Added party dependency, you can join with more players to one arena
 
 ### 4.3.1 Hotfix (13.12.2019)
 * Arena will not be stopped anymore on starting phase when there are null players (only resetting it to waiting for players)